Comentarios y modificacion de parametros evaluacion
enemigos y cofres en funcion del tamaño del grafo.
Hay que verlo.

+++ AG/AG/ParametrosEval.hpp
@@ -4,13 +4,13 @@
 
 class ParametrosEval {
 public:
-	unsigned int nodosOptimos;
-	double gradoOptimo;
-	unsigned int anchoOptimo;
-	unsigned int altoOptimo;
-	unsigned int ciclosOptimos;
-	unsigned int enemigosOptimos;
-	unsigned int cofresOptimos;
+	unsigned int nodosOptimos;		// Numero de nodos optimos
+	double gradoOptimo;				// Media de grado (aristas de cada nodo) optima
+	unsigned int anchoOptimo;		// Media de ancho de sala optima
+	unsigned int altoOptimo;		// Media de alto de sala optima
+	unsigned int ciclosOptimos;		// Numero de ciclos optimo
+	double enemigosOptimos;			// Densidad de enemigos (en funcion del numero de salas)
+	double cofresOptimos;			// Densidad de cofres (en funcion del numero de salas)
 };
 
 #endif
